requirement.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==2.0
idna==2.10
lxml==4.6.2
mysqlclient==1.4.2
Pillow==8.1.0
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.3
